requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
pydantic
python-dotenv
httpx
sqlparse
alembic
ollama
jinja2
mysql-connector-python